Refactored connected players layout
Exported planet images into pngs that are all 106px X 106px. Added folders "remote" and "display" to seperate components and redirected them in the AppDisplay.vue and AppRemote.vue files. Added a PlayerCard.vue component to remote folder that holds the styling and code to loop and display the players connected on the GameSession.vue page.